requirement.txt
mysqlclient-2.1.0
django >= 4.0.0, <4.1.0
djangorestframework
